Fixed the rendering; added fps limiting.
The rendering now works and is much faster. Haskell
could be made faster likely by passing an array of
pixels to be filled to Haskell and letting it fill
it, then returning it to C. It's probably the amount
of switches to Haskell code from C each frame that
causes the slowdown.

Also added basic FPS limiting. It would be better to
use a higher precision timing mechanism because SDL_Delay()
is only precise up to the millisecond level.